+++ src/examples.rs
@@ -55,8 +55,13 @@ local folder:
 {example8a}
 {example8b}
 
-`and` can be used to serve any number of routes simultaneously.
+Return HTTP status codes for some paths:
+{example9}
 
+Declare routes that do nothing using \"nothing\" (can be useful for scripted use):
+{example10}
+
+`and` can be used to serve any number of routes simultaneously.
 ",
     EXAMPLES="EXAMPLES:".bold(),
 
@@ -112,6 +117,13 @@ local folder:
 # http://localhost:8080/wibble/api/foo => ./files/foo.json
 # http://localhost:8080/bar/api/foo => ./files/foo.json
 # http://localhost:8080/api/foo => No route matches this".white()
+
+    example9="# Pick a specific status code (only valid HTTP status codes are allowed):
+weave 8080 to statuscode://403
+# The alias \"nothing\" returns a 404 Not Found status:
+weave 8080 to nothing"
+
+    example10="weave nothing and 8080 to 9090"
 
     ))
 }